Bumped the minimum Ruby dependency to 1.8.7 (or 1.8.1 with Backports).
Note that older Ruby 1.8.x versions are still supported with help of the Backports gem, but it is time to follow the lead of Rails, RDoc, RubyGems, etc, in ceasing to actively prop up Ruby 1.8.6 or yet older versions.

This means it is now permitted and safe to use Ruby 1.8.7+ idioms in the SXP.rb code base. In particular, note that on 1.8.7+:

* #each_* methods generally return an Enumerator when called without a block.
* Array#combination, #cycle, #permutation, #shuffle, #sample, #pop(n) and #shift(n) are available.
* Enumerable#first, #count(x), #none?, #one?, #take, #drop and #reverse_each are available.
* Enumerator#next and #rewind are available.
* Integer#even?, #odd? and #ord are available.
* IO#each_char, #chars, #lines, #bytes, #getbyte and #readbyte are available.
* Kernel#tap and #__method__ are available.
* Kernel#instance_exec, Module#module_exec and Module#class_exec are available.
* String#start_with? and #end_with? are available.
* String#each_char, #chars, #chr, #lines, #bytes, and #bytesize are available.
* Symbol#to_proc is available, enabling map(&:to_s) as shorthand for map{|s| s.to_s}.